[jobs] Fixed race condition with consecutive calls to Job::QueueService::wait()
A signal is used for  waiting for all worker threads of a
queueservice (the workers have to notify when they have finished their
work). During its lifetime, the queueservice may enter several times
into an idle state. Thus this signal must be reset, what the method
`wait` does if workers are still executing some job. However one
situation was not handled (where the signal must be reset too), when
the waiting room is not empty but the workers are really idle
(and about to be started).