String: the variadic methods `String::format()` and `String::appendFormat()` can not accept C++ classes as input
Those methods can not take C++ classes due to the intrinsic model of
vargs. 'va_start' has undefined behavior with reference types since it
is a macro which merely transform the following:

	va_start(valist, format)

into

	valist = (va_list) (&format + 1)

If format is a value-type, it gets placed on the stack right before all
the variadic arguments. If format is a reference type, only the address
gets placed on the stack. When you take the address of the reference
variable, you get the address or the original variable (in this case of
a temporary AnsiString created before calling Broken), not the address
of the argument