Yuni::async: fixed a race condition when creating the new job
The job is created using a weak pointer. It is added then to the queue
service and then returned. However, it may be possible that the job is
done before even returning from the function (and thus the job is
already destroyed by the queueservice since no one has acquired a
reference on it)